[CHORE 컴포넌트 폴더 정리, IMG, SVG분리
